move script_utils to schnetpack
