+++ src/schnetpack/utils/__init__.py
@@ -1,0 +1,1 @@
+from .spk_utils import *
