+++ src/schnetpack/utils/script_utils/data.py
