+++ src/schnetpack/utils/script_utils/model.py
