+++ src/scripts/schnetpack_qm9.py
@@ -2,18 +2,17 @@
 import logging
 import os
 
-import schnetpack.output_modules
 import torch
 from ase.data import atomic_numbers
 
 import schnetpack as spk
 from schnetpack.datasets import QM9
-from scripts.script_utils.script_parsing import get_main_parser, add_subparsers
-from scripts.script_utils.setup import setup_run
-from scripts.script_utils.model import get_representation, get_model
-from scripts.script_utils.training import train
-from scripts.script_utils.evaluation import evaluate
-from scripts.script_utils import get_statistics, get_loaders
+from schnetpack.utils.script_utils import get_main_parser, add_subparsers
+from schnetpack.utils.script_utils.setup import setup_run
+from schnetpack.utils.script_utils import get_representation, get_model
+from schnetpack.utils.script_utils.training import train
+from schnetpack.utils.script_utils.evaluation import evaluate
+from schnetpack.utils.script_utils import get_statistics, get_loaders
 
 logging.basicConfig(level=os.environ.get("LOGLEVEL", "INFO"))
